Refactor project by renaming Blink3.Common to Blink3.Core
The formerly known Blink3.Common project has been refactored to Blink3.Core for better semantic meaning and organization. All the namespaces and references have been updated accordingly. Files have also been moved to reflect this change.
