+++ Blink3.API/Blink3.API.csproj
@@ -36,7 +36,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Blink3.Common\Blink3.Common.csproj"/>
+        <ProjectReference Include="..\Blink3.Core\Blink3.Core.csproj"/>
         <ProjectReference Include="..\Blink3.DataAccess\Blink3.DataAccess.csproj"/>
     </ItemGroup>
