+++ Blink3.API/Controllers/ApiControllerBase.cs
@@ -1,5 +1,5 @@
 using System.Net.Mime;
-using Blink3.Common.Extensions;
+using Blink3.Core.DiscordAuth.Extensions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
