+++ Blink3.Bot/Blink3.Bot.csproj
@@ -24,7 +24,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Blink3.Common\Blink3.Common.csproj"/>
+        <ProjectReference Include="..\Blink3.Core\Blink3.Core.csproj"/>
         <ProjectReference Include="..\Blink3.DataAccess\Blink3.DataAccess.csproj"/>
     </ItemGroup>
