+++ Blink3.Bot/Program.cs
@@ -1,7 +1,7 @@
 ﻿using Blink3.Bot.Services;
-using Blink3.Common.Caching.Extensions;
-using Blink3.Common.Configuration;
-using Blink3.Common.Configuration.Extensions;
+using Blink3.Core.Caching.Extensions;
+using Blink3.Core.Configuration;
+using Blink3.Core.Configuration.Extensions;
 using Blink3.DataAccess.Extensions;
 using Blink3.DataAccess.Services;
 using Discord;
