+++ Blink3.Core/Blink3.Core.csproj
@@ -18,4 +18,8 @@
         <PackageReference Include="Microsoft.Extensions.Options.DataAnnotations" Version="8.0.0"/>
     </ItemGroup>
 
+    <ItemGroup>
+        <Folder Include="Repositories\"/>
+    </ItemGroup>
+
 </Project>
