+++ Blink3.Core/Repositories/Interfaces/IBlinkGuildRepository.cs
