+++ Blink3.Core/Repositories/Interfaces/IGenericRepository.cs
