+++ Blink3.DataAccess/Extensions/ServiceCollectionExtensions.cs
@@ -1,4 +1,4 @@
-using Blink3.Common.Configuration;
+using Blink3.Core.Configuration;
 using Blink3.DataAccess.Interfaces;
 using Blink3.DataAccess.Repositories;
 using Microsoft.EntityFrameworkCore;
