+++ Blink3.Web/Blink3.Web.csproj
@@ -26,8 +26,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Blink3.Common\Blink3.Common.csproj"/>
-        <ProjectReference Include="..\Blink3.DataAccess\Blink3.DataAccess.csproj"/>
+        <ProjectReference Include="..\Blink3.Core\Blink3.Core.csproj"/>
     </ItemGroup>
 
     <ItemGroup>
