+++ Blink3.Web/Components/LoginDisplay.razor
@@ -1,4 +1,4 @@
-@using Blink3.Common.Extensions
+@using Blink3.Core.DiscordAuth.Extensions
 @inject NavigationManager Navigation
 
 <AuthorizeView>
