+++ Blink3.sln.DotSettings
@@ -1,4 +1,5 @@
-﻿<wpf:ResourceDictionary xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:s="clr-namespace:System;assembly=mscorlib"
+﻿<wpf:ResourceDictionary xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                        xmlns:s="clr-namespace:System;assembly=mscorlib"
                         xmlns:wpf="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                         xml:space="preserve">
 	<s:Boolean x:Key="/Default/UserDictionary/Words/=lavalink/@EntryIndexedValue">True</s:Boolean></wpf:ResourceDictionary>
